// src/app/enums/cloth-color.enum.ts
export enum ClothColor {
  BLACK = 'black',
  BLUE1 = 'blue1',
  BLUE2 = 'blue2',
  BLUE3 = 'blue3',
  GRAY1 = 'gray1',
  GRAY2 = 'gray2',
  HEATHER = 'heather',
  PASTEL_BLUE = 'pastel_blue',
  PASTEL_GREEN = 'pastel_green',
  PASTEL_ORANGE = 'pastel_orange',
  PASTEL_RED = 'pastel_red', 
  PASTEL_YELLOW = 'pastel_yellow',
  PINK = 'pink',
  RED = 'red',
  WHITE = 'white'
}