// src/app/enums/hat-color.enum.ts
export enum HatColor {
  BLACK = 'black',
  BLUE01 = 'blue01',
  BLUE02  = 'blue02',
  BLUE03 = 'blue03',
  GRAY01 = 'gray01',
  GRAY02 = 'gray02',
  HEATHER = 'heather',
  PASTEL_BLUE = 'pastel_blue',
  PASTEL_GREEN = 'pastel_green',
  PASTEL_ORANGE = 'pastel_orange',
  PASTEL_RED = 'pastel_red',
  PASTEL_YELLOW = 'pastel_yellow',
  PINK = 'pink',
  RED = 'red',
  WHITE = 'white',
}